fix: resolve correctness issues across multiple packages
- Fix CUC FineTime int64 overflow for extended fine bytes (>=5) by
  switching to uint64 and successive-doubling arithmetic
- Add DecodeTCTransferFrameWithSegmentHeader() to properly parse MAP
  segment headers on decode, fixing roundtrip asymmetry
- Add input validation to BCHEncode to prevent panic on short slices
- Decouple VCP receive gap resync from send-side FrameCounter via
  SetGapResync() so pure receivers can detect frame loss
- Fix misleading evalPoly comment (direct evaluation, not Horner's)

+++ cli/time.go
@@ -40,7 +40,7 @@ type timeJSON struct {
 
 	// CUC fields
 	CoarseTime  *uint64 `json:"coarse_time,omitempty"`
-	FineTime    *uint32 `json:"fine_time,omitempty"`
+	FineTime    *uint64 `json:"fine_time,omitempty"`
 	CoarseBytes *uint8  `json:"coarse_bytes,omitempty"`
 	FineBytes   *uint8  `json:"fine_bytes,omitempty"`
